feat: correction config live rescan, adaptive timeline, UX fixes
- Fix correction config not applied on re-analysis (read from
  correction_customizations instead of legacy correction_config key)
- Selective typography re-scan on config save: alerts update instantly
  without full re-analysis (~100ms regex-only detector)
- Frontend refreshes alerts after config save (@saved → loadAlerts)
- Timeline: adaptive axis labels (Día X for days, +Xh for hours)
- Timeline: sub-day granularity (day_offset now float, preserves hours)
- Timeline: adaptive zoomMin/padding based on story time span
- Timeline: tooltip z-index fix (grid lines no longer overlap popover)
- Chat error messages: user-friendly (point to Ajustes > LLM)
- Bump version to 0.9.10

+++ src/narrative_assistant/persistence/timeline.py
@@ -29,7 +29,7 @@ class TimelineEventData:
     discourse_position: int | None = None
     confidence: float = 0.5
     # Para timelines sin fechas absolutas (Día 0, Día +1, etc.)
-    day_offset: int | None = None
+    day_offset: float | None = None
     weekday: str | None = None
     # Instancia temporal (viajes en el tiempo: A@40 vs A@45)
     temporal_instance_id: str | None = None
